fix(quality): resolve codacy integer-overflow in WikiPrice.midPrice
Widen the sum to long before averaging so instaBuy + instaSell can't
overflow int for high-value items.

+++ src/main/java/com/flipsmart/api/dto/WikiPrice.java
@@ -37,6 +37,6 @@ public class WikiPrice
 		{
 			return instaBuy;
 		}
-		return (instaBuy + instaSell) / 2;
+		return (int) (((long) instaBuy + instaSell) / 2);
 	}
 }
